Extract NTP printing
This method is required to disect some RADIUS attributes
