+++ Makefile.in
@@ -87,6 +87,7 @@ LIBNETDISSECT_SRC=\
 	netdissect.c \
 	netdissect-alloc.c \
 	nlpid.c \
+	ntp.c \
 	oui.c \
 	parsenfsfh.c \
 	print.c \
@@ -286,6 +287,7 @@ HDR = \
 	nfs.h \
 	nfsfh.h \
 	nlpid.h \
+	ntp.h \
 	openflow.h \
 	ospf.h \
 	oui.h \
